+++ apps/lp/components/ui/header.tsx
@@ -1,5 +1,9 @@
+'use client';
 import Image from 'next/image';
 import Link from 'next/link';
+import React from 'react';
+
+import LinkWithUTMFromPath from '@chaindesk/ui/LinkWithUTMFromPath';
 
 import Menu from './products-menu';
 
@@ -12,9 +16,9 @@ export default function Header() {
         <div className="mx-auto max-w-4xl">
           <div className="flex items-center justify-between h-14 border border-transparent [background:linear-gradient(theme(colors.white),theme(colors.white))_padding-box,linear-gradient(120deg,theme(colors.zinc.300),theme(colors.zinc.100),theme(colors.zinc.300))_border-box] rounded-lg px-3">
             {/* Site branding */}
-            <Link
+            <LinkWithUTMFromPath
               className="inline-flex items-center mr-4 space-x-2 shrink-0"
-              href="/"
+              href={`/`}
             >
               {/* Logo */}
               <div className="flex justify-center items-center w-8 h-8 bg-white rounded shadow-sm shadow-zinc-950/20">
@@ -23,7 +27,7 @@ export default function Header() {
               <span className="text-lg font-bold font-bricolage-grotesque">
                 Chaindesk
               </span>
-            </Link>
+            </LinkWithUTMFromPath>
 
             {/* Desktop navigation */}
             <nav className="flex grow">
@@ -61,21 +65,21 @@ export default function Header() {
                 </li>
 
                 <li>
-                  <Link
+                  <LinkWithUTMFromPath
                     className="hidden items-center px-3 py-2 text-sm font-medium transition md:flex text-zinc-500 hover:text-zinc-900 lg:px-5"
-                    href="https://app.chaindesk.ai/signin"
+                    href={`https://app.chaindesk.ai/signin`}
                   >
                     Log in
-                  </Link>
+                  </LinkWithUTMFromPath>
                 </li>
 
                 <li className="ml-1">
-                  <Link
+                  <LinkWithUTMFromPath
                     className="hidden w-full shadow sm:flex btn-sm text-zinc-100 bg-zinc-900 hover:bg-zinc-800"
-                    href="https://app.chaindesk.ai/signin"
+                    href={`https://app.chaindesk.ai/signin`}
                   >
                     Sign up for free
-                  </Link>
+                  </LinkWithUTMFromPath>
                 </li>
               </ul>
             </nav>